rename Spree::Globalize to SpreeGlobalize, closes #16
