+++ app/controllers/spree/admin/general_settings_controller_decorator.rb
@@ -4,8 +4,8 @@ Spree::Admin::GeneralSettingsController.class_eval do
   private
     def update_i18n_settings
       params.each do |name, value|
-        next unless Spree::Globalize::Config.has_preference? name
-        Spree::Globalize::Config[name] = value.map(&:to_sym)
+        next unless SpreeGlobalize::Config.has_preference? name
+        SpreeGlobalize::Config[name] = value.map(&:to_sym)
       end
     end
 end
