+++ app/controllers/spree/base_controller_decorator.rb
@@ -1,3 +1,3 @@
 Spree::BaseController.class_eval do
-  include Spree::Globalize::ControllerGlobalizeHelper
+  include SpreeGlobalize::ControllerGlobalizeHelper
 end
