+++ app/models/concerns/spree_globalize/translatable.rb
@@ -1,0 +1,34 @@
+module SpreeGlobalize
+  module Translatable
+    extend ActiveSupport::Concern
+
+    included do |klass|
+      accepts_nested_attributes_for :translations
+      klass.whitelisted_ransackable_associations |= ['translations']
+    end
+
+    class_methods do
+      def ransack(params = {}, options = {})
+        params ||= {}
+        names = params.keys
+
+        names.each do |n|
+          translated_attribute_names.each do |t|
+            if n.to_s.starts_with? t.to_s
+              params[:"translations_#{n}"] = params[n]
+              params.delete n
+            end
+          end
+        end
+
+        super(params, options)
+      end
+      alias :search :ransack unless respond_to? :search
+
+      # preload translations
+      def spree_base_scopes
+        super.includes(:translations).references(:translations)
+      end
+    end
+  end
+end
