+++ app/models/spree/option_type_decorator.rb
@@ -1,6 +1,6 @@
 module Spree
   OptionType.class_eval do
     translates :name, :presentation, fallbacks_for_empty_translations: true
-    include Spree::Globalize::Translatable
+    include SpreeGlobalize::Translatable
   end
 end
