+++ app/models/spree/product_property_decorator.rb
@@ -1,6 +1,6 @@
 module Spree
   ProductProperty.class_eval do
     translates :value, fallbacks_for_empty_translations: true
-    include Spree::Globalize::Translatable
+    include SpreeGlobalize::Translatable
   end
 end
