+++ app/models/spree/promotion_decorator.rb
@@ -1,6 +1,6 @@
 module Spree
   Promotion.class_eval do
     translates :name, :description, fallbacks_for_empty_translations: true
-    include Spree::Globalize::Translatable
+    include SpreeGlobalize::Translatable
   end
 end
