+++ app/models/spree/shipping_method_decorator.rb
@@ -1,6 +1,6 @@
 module Spree
   ShippingMethod.class_eval do
     translates :name, fallbacks_for_empty_translations: true
-    include Spree::Globalize::Translatable
+    include SpreeGlobalize::Translatable
   end
 end
