+++ app/models/spree/taxon_decorator.rb
@@ -2,6 +2,6 @@ module Spree
   Taxon.class_eval do
     translates :name, :description, :meta_title, :meta_description, :meta_keywords,
       :permalink, fallbacks_for_empty_translations: true
-    include Spree::Globalize::Translatable
+    include SpreeGlobalize::Translatable
   end
 end
