+++ app/overrides/spree/shared/_main_nav_bar/locale_selector.html.erb.deface
@@ -1,5 +1,5 @@
 <!-- insert_bottom '#main-nav-bar > .navbar-right' -->
-<% if Spree::Globalize::Config.supported_locales.many? %>
+<% if SpreeGlobalize::Config.supported_locales.many? %>
   <li id="locale-select" data-hook>
     <%= form_tag spree.set_locale_path, class: 'navbar-form' do %>
       <div class="form-group">
