--- lib/spree/globalize/controller_globalize_helper.rb
@@ -1,20 +1,0 @@
-module Spree
-  module Globalize
-    # The fact this logic is in a single module also helps to apply a custom
-    # locale on the spree/api context since api base controller inherits from
-    # MetalController instead of Spree::BaseController
-    module ControllerGlobalizeHelper
-      extend ActiveSupport::Concern
-
-      included do
-        prepend_before_filter :globalize_fallbacks
-
-        private
-
-        def globalize_fallbacks
-          Fallbacks.config!
-        end
-      end
-    end
-  end
-end
