--- lib/spree/globalize/version.rb
@@ -1,20 +1,0 @@
-module Spree
-  module Globalize
-    module_function
-
-    # Returns the version of the currently loaded Spree::Globalize as a
-    # <tt>Gem::Version</tt>.
-    def version
-      Gem::Version.new VERSION::STRING
-    end
-
-    module VERSION
-      MAJOR = 3
-      MINOR = 1
-      TINY  = 0
-      PRE   = "beta"
-
-      STRING = [MAJOR, MINOR, TINY, PRE].compact.join('.')
-    end
-  end
-end
