+++ lib/spree_globalize.rb
@@ -1,5 +1,5 @@
 require 'sass/rails'
 require 'spree_i18n'
-require 'spree/globalize/engine'
-require 'spree/globalize/version'
+require 'spree_globalize/engine'
+require 'spree_globalize/version'
 require 'coffee_script'
