+++ spec/features/admin/translations_spec.rb
@@ -8,7 +8,7 @@ RSpec.feature "Translations", :js do
     create(:store)
     reset_spree_preferences
     SpreeI18n::Config.available_locales = [:en, :'pt-BR']
-    Spree::Globalize::Config.supported_locales = [:en, :'pt-BR']
+    SpreeGlobalize::Config.supported_locales = [:en, :'pt-BR']
   end
 
   context "products" do
@@ -230,7 +230,7 @@ RSpec.feature "Translations", :js do
     scenario "adds german to supported locales" do
       targetted_select2_search(language, from: '#s2id_supported_locales_')
       click_on 'Update'
-      expect(Spree::Globalize::Config.supported_locales).to include(:de)
+      expect(SpreeGlobalize::Config.supported_locales).to include(:de)
     end
   end
 
@@ -240,7 +240,7 @@ RSpec.feature "Translations", :js do
 
     scenario "finds the right product with permalink in a not active language" do
       SpreeI18n::Config.available_locales = [:en, :de]
-      Spree::Globalize::Config.supported_locales = [:en, :de]
+      SpreeGlobalize::Config.supported_locales = [:en, :de]
 
       visit spree.admin_product_path(product)
       click_on "Translations"
