+++ spec/features/translations_spec.rb
@@ -3,7 +3,7 @@ RSpec.feature "Translations" do
   background do
     reset_spree_preferences
     SpreeI18n::Config.available_locales = [:en, :'pt-BR']
-    Spree::Globalize::Config.supported_locales = [:en, :'pt-BR']
+    SpreeGlobalize::Config.supported_locales = [:en, :'pt-BR']
   end
 
   context 'product' do
